# docs/openapi.json
{"openapi":"3.0.0","paths":{"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"Usuário criado; token JWT no corpo da resposta"},"409":{"description":"E-mail ou CPF já cadastrado"}},"summary":"Cria um usuário e já devolve o token de acesso","tags":["Auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Token JWT + dados do usuário"},"401":{"description":"Credenciais inválidas"}},"summary":"Autentica por e-mail e senha e devolve o token JWT","tags":["Auth"]}},"/api/auth/perfil":{"get":{"operationId":"AuthController_perfil","parameters":[],"responses":{"200":{"description":"Usuário do token"}},"security":[{"bearer":[]}],"summary":"Dados do usuário autenticado (token atual)","tags":["Auth"]}},"/api/usuarios":{"post":{"operationId":"UsuariosController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUsuarioDto"}}}},"responses":{"201":{"description":"Usuário criado"},"409":{"description":"E-mail ou CPF já cadastrado"}},"security":[{"bearer":[]}],"summary":"Cria um usuário","tags":["Usuários"]},"get":{"operationId":"UsuariosController_findAll","parameters":[{"name":"pagina","required":false,"in":"query","description":"Página (a partir de 1)","schema":{"default":1,"example":1,"type":"number"}},{"name":"limite","required":false,"in":"query","description":"Itens por página (máx. 100)","schema":{"default":10,"example":10,"type":"number"}},{"name":"busca","required":false,"in":"query","description":"Busca textual livre","schema":{"example":"Recife","type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Lista usuários com paginação e busca por nome/e-mail/CPF","tags":["Usuários"]}},"/api/usuarios/{id}":{"get":{"operationId":"UsuariosController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"404":{"description":"Usuário não encontrado"}},"security":[{"bearer":[]}],"summary":"Busca um usuário pelo ID","tags":["Usuários"]},"patch":{"operationId":"UsuariosController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUsuarioDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Atualiza parcialmente um usuário","tags":["Usuários"]},"delete":{"operationId":"UsuariosController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":""}},"security":[{"bearer":[]}],"summary":"Exclui um usuário (soft delete — preserva o histórico)","tags":["Usuários"]}},"/api/usuarios/{id}/restaurar":{"patch":{"operationId":"UsuariosController_restore","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Restaura um usuário excluído","tags":["Usuários"]}},"/api/cartorios":{"post":{"operationId":"CartoriosController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCartorioDto"}}}},"responses":{"201":{"description":"Cartório criado"},"409":{"description":"CNPJ já cadastrado"}},"security":[{"bearer":[]}],"summary":"Cria um cartório","tags":["Cartórios"]},"get":{"operationId":"CartoriosController_findAll","parameters":[{"name":"pagina","required":false,"in":"query","description":"Página (a partir de 1)","schema":{"default":1,"example":1,"type":"number"}},{"name":"limite","required":false,"in":"query","description":"Itens por página (máx. 100)","schema":{"default":10,"example":10,"type":"number"}},{"name":"busca","required":false,"in":"query","description":"Busca textual livre","schema":{"example":"Recife","type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Lista cartórios com paginação e busca por nome/CNPJ/cidade","tags":["Cartórios"]}},"/api/cartorios/{id}":{"get":{"operationId":"CartoriosController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"404":{"description":"Cartório não encontrado"}},"security":[{"bearer":[]}],"summary":"Busca um cartório pelo ID","tags":["Cartórios"]},"patch":{"operationId":"CartoriosController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCartorioDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Atualiza parcialmente um cartório","tags":["Cartórios"]},"delete":{"operationId":"CartoriosController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"409":{"description":"Cartório possui usuários ou imóveis ativos"}},"security":[{"bearer":[]}],"summary":"Exclui um cartório (soft delete; bloqueado se houver vínculos ativos)","tags":["Cartórios"]}},"/api/cartorios/{id}/restaurar":{"patch":{"operationId":"CartoriosController_restore","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Restaura um cartório excluído","tags":["Cartórios"]}},"/api/imoveis":{"post":{"operationId":"ImoveisController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateImovelDto"}}}},"responses":{"201":{"description":"Imóvel criado"},"404":{"description":"Cartório informado não existe"},"409":{"description":"Matrícula já registrada neste cartório"}},"security":[{"bearer":[]}],"summary":"Registra um imóvel em um cartório","tags":["Imóveis"]},"get":{"operationId":"ImoveisController_findAll","parameters":[{"name":"pagina","required":false,"in":"query","description":"Página (a partir de 1)","schema":{"default":1,"example":1,"type":"number"}},{"name":"limite","required":false,"in":"query","description":"Itens por página (máx. 100)","schema":{"default":10,"example":10,"type":"number"}},{"name":"busca","required":false,"in":"query","description":"Busca textual livre","schema":{"example":"Recife","type":"string"}},{"name":"tipo","required":false,"in":"query","description":"Filtra por tipo","schema":{"type":"string","enum":["casa","apartamento","terreno","sala_comercial","galpao","rural","outro"]}},{"name":"status","required":false,"in":"query","description":"Filtra por status","schema":{"type":"string","enum":["regular","pendente","alienado","bloqueado","cancelado"]}},{"name":"cartorio_id","required":false,"in":"query","description":"Filtra por cartório","schema":{"example":1,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Lista imóveis com paginação, busca (matrícula/cidade/proprietário) e filtros por tipo, status e cartório","tags":["Imóveis"]}},"/api/imoveis/{id}":{"get":{"operationId":"ImoveisController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"404":{"description":"Imóvel não encontrado"}},"security":[{"bearer":[]}],"summary":"Busca um imóvel pelo ID","tags":["Imóveis"]},"patch":{"operationId":"ImoveisController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateImovelDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Atualiza parcialmente um imóvel","tags":["Imóveis"]},"delete":{"operationId":"ImoveisController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":""}},"security":[{"bearer":[]}],"summary":"Exclui um imóvel (soft delete — preserva o histórico)","tags":["Imóveis"]}},"/api/imoveis/{id}/restaurar":{"patch":{"operationId":"ImoveisController_restore","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Restaura um imóvel excluído","tags":["Imóveis"]}},"/api/relatorios/resumo":{"get":{"operationId":"RelatoriosController_resumo","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Totais gerais: cartórios, usuários, imóveis, valor e área","tags":["Relatórios"]}},"/api/relatorios/imoveis-por-cartorio":{"get":{"operationId":"RelatoriosController_imoveisPorCartorio","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Imóveis, usuários e valor avaliado consolidados por cartório","tags":["Relatórios"]}},"/api/relatorios/imoveis-por-tipo":{"get":{"operationId":"RelatoriosController_imoveisPorTipo","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Distribuição de imóveis por tipo","tags":["Relatórios"]}},"/api/relatorios/imoveis-por-status":{"get":{"operationId":"RelatoriosController_imoveisPorStatus","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Distribuição de imóveis por status","tags":["Relatórios"]}}},"info":{"title":"Desafio Sião — API de Cartórios e Imóveis","description":"API CRUD de cartórios, usuários e imóveis com autenticação JWT.\n\n**Como autenticar:** crie um usuário em `POST /api/auth/register` (ou use o usuário do seed: `admin@siao.com.br` / `Admin@123`), faça login em `POST /api/auth/login`, copie o `access_token` e clique em **Authorize** aqui no Swagger.","version":"1.0.0","contact":{}},"tags":[{"name":"Auth","description":"Registro, login e perfil do usuário autenticado"},{"name":"Cartórios","description":"CRUD de cartórios"},{"name":"Usuários","description":"CRUD de usuários"},{"name":"Imóveis","description":"CRUD de imóveis"},{"name":"Relatórios","description":"Consolidados e agregações"}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"nome":{"type":"string","example":"Maria da Silva","maxLength":120},"cpf":{"type":"string","example":"111.444.777-35","description":"CPF válido, com ou sem máscara"},"email":{"type":"string","example":"maria@exemplo.com.br","maxLength":160},"password":{"type":"string","example":"Senha@123","minLength":8,"description":"Mínimo 8 caracteres, com ao menos uma letra e um número"},"telefone":{"type":"string","example":"(81) 98888-7777","maxLength":20},"endereco":{"type":"string","example":"Rua do Sol, 120, ap 301","maxLength":200},"cidade":{"type":"string","example":"Recife","maxLength":80},"estado":{"type":"string","example":"PE","description":"UF com 2 letras maiúsculas"},"cep":{"type":"string","example":"50030-230","maxLength":9},"cartorio_id":{"type":"number","example":1,"description":"ID do cartório ao qual o usuário pertence"}},"required":["nome","cpf","email","password","telefone","endereco","cidade","estado","cep"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"admin@siao.com.br"},"password":{"type":"string","example":"Admin@123","minLength":8}},"required":["email","password"]},"CreateUsuarioDto":{"type":"object","properties":{"nome":{"type":"string","example":"Maria da Silva","maxLength":120},"cpf":{"type":"string","example":"111.444.777-35","description":"CPF válido, com ou sem máscara"},"email":{"type":"string","example":"maria@exemplo.com.br","maxLength":160},"password":{"type":"string","example":"Senha@123","minLength":8,"description":"Mínimo 8 caracteres, com ao menos uma letra e um número"},"telefone":{"type":"string","example":"(81) 98888-7777","maxLength":20},"endereco":{"type":"string","example":"Rua do Sol, 120, ap 301","maxLength":200},"cidade":{"type":"string","example":"Recife","maxLength":80},"estado":{"type":"string","example":"PE","description":"UF com 2 letras maiúsculas"},"cep":{"type":"string","example":"50030-230","maxLength":9},"cartorio_id":{"type":"number","example":1,"description":"ID do cartório ao qual o usuário pertence"}},"required":["nome","cpf","email","password","telefone","endereco","cidade","estado","cep"]},"UpdateUsuarioDto":{"type":"object","properties":{"nome":{"type":"string","example":"Maria da Silva","maxLength":120},"cpf":{"type":"string","example":"111.444.777-35","description":"CPF válido, com ou sem máscara"},"email":{"type":"string","example":"maria@exemplo.com.br","maxLength":160},"password":{"type":"string","example":"Senha@123","minLength":8,"description":"Mínimo 8 caracteres, com ao menos uma letra e um número"},"telefone":{"type":"string","example":"(81) 98888-7777","maxLength":20},"endereco":{"type":"string","example":"Rua do Sol, 120, ap 301","maxLength":200},"cidade":{"type":"string","example":"Recife","maxLength":80},"estado":{"type":"string","example":"PE","description":"UF com 2 letras maiúsculas"},"cep":{"type":"string","example":"50030-230","maxLength":9},"cartorio_id":{"type":"number","example":1,"description":"ID do cartório ao qual o usuário pertence"}}},"CreateCartorioDto":{"type":"object","properties":{"nome":{"type":"string","example":"1º Cartório de Registro de Imóveis de Recife","maxLength":160},"cnpj":{"type":"string","example":"11.222.333/0001-81","description":"CNPJ válido, com ou sem máscara"},"telefone":{"type":"string","example":"(81) 3222-1100","maxLength":20},"email":{"type":"string","example":"contato@cartoriorecife.com.br","maxLength":160},"logradouro":{"type":"string","example":"Avenida Rio Branco","maxLength":160},"numero":{"type":"number","example":240},"bairro":{"type":"string","example":"Recife Antigo","maxLength":80},"cidade":{"type":"string","example":"Recife","maxLength":80},"estado":{"type":"string","example":"PE","description":"UF com 2 letras maiúsculas"},"cep":{"type":"string","example":"50030-310","maxLength":9},"responsavel_id":{"type":"number","example":10,"description":"Identificador externo do responsável"},"responsavel_nome":{"type":"string","example":"João Pereira Lima","maxLength":120},"responsavel_cpf":{"type":"string","example":"111.444.777-35","description":"CPF válido do responsável"}},"required":["nome","cnpj","telefone","email","logradouro","numero","bairro","cidade","estado","cep","responsavel_nome","responsavel_cpf"]},"UpdateCartorioDto":{"type":"object","properties":{"nome":{"type":"string","example":"1º Cartório de Registro de Imóveis de Recife","maxLength":160},"cnpj":{"type":"string","example":"11.222.333/0001-81","description":"CNPJ válido, com ou sem máscara"},"telefone":{"type":"string","example":"(81) 3222-1100","maxLength":20},"email":{"type":"string","example":"contato@cartoriorecife.com.br","maxLength":160},"logradouro":{"type":"string","example":"Avenida Rio Branco","maxLength":160},"numero":{"type":"number","example":240},"bairro":{"type":"string","example":"Recife Antigo","maxLength":80},"cidade":{"type":"string","example":"Recife","maxLength":80},"estado":{"type":"string","example":"PE","description":"UF com 2 letras maiúsculas"},"cep":{"type":"string","example":"50030-310","maxLength":9},"responsavel_id":{"type":"number","example":10,"description":"Identificador externo do responsável"},"responsavel_nome":{"type":"string","example":"João Pereira Lima","maxLength":120},"responsavel_cpf":{"type":"string","example":"111.444.777-35","description":"CPF válido do responsável"}}},"CreateImovelDto":{"type":"object","properties":{"matricula":{"type":"string","example":"MAT-2026-000123","maxLength":40,"description":"Matrícula do imóvel (única por cartório)"},"tipo":{"type":"string","enum":["casa","apartamento","terreno","sala_comercial","galpao","rural","outro"],"example":"apartamento"},"logradouro":{"type":"string","example":"Rua da Aurora","maxLength":160},"numero":{"type":"number","example":1520},"bairro":{"type":"string","example":"Boa Vista","maxLength":80},"cidade":{"type":"string","example":"Recife","maxLength":80},"estado":{"type":"string","example":"PE","description":"UF com 2 letras maiúsculas"},"cep":{"type":"string","example":"50050-000","maxLength":9},"area_total":{"type":"number","example":78.5,"description":"Área total em m² (maior que zero)"},"valor_avaliado":{"type":"number","example":420000,"description":"Valor de avaliação em reais"},"status":{"type":"string","enum":["regular","pendente","alienado","bloqueado","cancelado"],"default":"regular"},"proprietario_id":{"type":"number","example":55,"description":"Identificador externo do proprietário"},"proprietario_nome":{"type":"string","example":"Ana Beatriz Costa","maxLength":120},"proprietario_cpf":{"type":"string","example":"111.444.777-35","description":"CPF válido do proprietário"},"cartorio_id":{"type":"number","example":1,"description":"ID do cartório onde o imóvel está registrado"}},"required":["matricula","tipo","logradouro","numero","bairro","cidade","estado","cep","area_total","valor_avaliado","proprietario_nome","proprietario_cpf","cartorio_id"]},"UpdateImovelDto":{"type":"object","properties":{"matricula":{"type":"string","example":"MAT-2026-000123","maxLength":40,"description":"Matrícula do imóvel (única por cartório)"},"tipo":{"type":"string","enum":["casa","apartamento","terreno","sala_comercial","galpao","rural","outro"],"example":"apartamento"},"logradouro":{"type":"string","example":"Rua da Aurora","maxLength":160},"numero":{"type":"number","example":1520},"bairro":{"type":"string","example":"Boa Vista","maxLength":80},"cidade":{"type":"string","example":"Recife","maxLength":80},"estado":{"type":"string","example":"PE","description":"UF com 2 letras maiúsculas"},"cep":{"type":"string","example":"50050-000","maxLength":9},"area_total":{"type":"number","example":78.5,"description":"Área total em m² (maior que zero)"},"valor_avaliado":{"type":"number","example":420000,"description":"Valor de avaliação em reais"},"status":{"type":"string","enum":["regular","pendente","alienado","bloqueado","cancelado"],"default":"regular"},"proprietario_id":{"type":"number","example":55,"description":"Identificador externo do proprietário"},"proprietario_nome":{"type":"string","example":"Ana Beatriz Costa","maxLength":120},"proprietario_cpf":{"type":"string","example":"111.444.777-35","description":"CPF válido do proprietário"},"cartorio_id":{"type":"number","example":1,"description":"ID do cartório onde o imóvel está registrado"}}}}}}